HAAR-889: Moved endpoint form external to generic
