+++ src/main/kotlin/uk/gov/justice/digital/hmpps/manageusersapi/resource/external/UserController.kt
@@ -161,31 +161,6 @@ class UserController(
     userId,
     deactivateReason
   )
-
-  @GetMapping("/externalusers/me/roles")
-  @Operation(
-    summary = "List of roles for current user.",
-    description = "List of roles for current user."
-  )
-  @ApiResponses(
-    value = [
-      ApiResponse(
-        responseCode = "200",
-        description = "OK"
-      ),
-      ApiResponse(
-        responseCode = "401",
-        description = "Unauthorized.",
-        content = [
-          Content(
-            mediaType = "application/json",
-            schema = Schema(implementation = ErrorResponse::class)
-          )
-        ]
-      )
-    ]
-  )
-  fun myRoles() = userService.myRoles()
 }
 
 @Schema(description = "Deactivate Reason")
