+++ src/test/kotlin/uk/gov/justice/digital/hmpps/manageusersapi/service/external/UserServiceTest.kt
@@ -1,8 +1,6 @@
 package uk.gov.justice.digital.hmpps.manageusersapi.service.external
 
 import com.microsoft.applicationinsights.TelemetryClient
-import kotlinx.coroutines.runBlocking
-import org.assertj.core.api.Assertions.assertThat
 import org.junit.jupiter.api.Nested
 import org.junit.jupiter.api.Test
 import org.mockito.kotlin.anyOrNull
@@ -10,10 +8,7 @@ import org.mockito.kotlin.mock
 import org.mockito.kotlin.verify
 import org.mockito.kotlin.verifyNoInteractions
 import org.mockito.kotlin.whenever
-import org.springframework.security.core.Authentication
-import org.springframework.security.core.authority.SimpleGrantedAuthority
 import uk.gov.justice.digital.hmpps.manageusersapi.adapter.external.UserApiService
-import uk.gov.justice.digital.hmpps.manageusersapi.config.AuthenticationFacade
 import uk.gov.justice.digital.hmpps.manageusersapi.resource.external.DeactivateReason
 import uk.gov.justice.digital.hmpps.manageusersapi.service.EmailNotificationService
 import java.util.UUID
@@ -22,14 +17,11 @@ class UserServiceTest {
   private val userApiService: UserApiService = mock()
   private val emailNotificationService: EmailNotificationService = mock()
   private val telemetryClient: TelemetryClient = mock()
-  private val authenticationFacade: AuthenticationFacade = mock()
-  private val authentication: Authentication = mock()
 
   private val userService = UserService(
     userApiService,
     emailNotificationService,
-    telemetryClient,
-    authenticationFacade,
+    telemetryClient
   )
   private val userUUID: UUID = UUID.fromString("00000000-aaaa-0000-aaaa-0a0a0a0a0a0a")
 
@@ -61,23 +53,6 @@ class UserServiceTest {
       val reason = DeactivateReason("Fired")
       userService.disableUserByUserId(userUUID, reason)
       verify(userApiService).disableUserById(UUID.fromString("00000000-aaaa-0000-aaaa-0a0a0a0a0a0a"), reason)
-    }
-  }
-
-  @Nested
-  inner class MyRoles {
-    @Test
-    fun myRoles(): Unit = runBlocking {
-      whenever(authenticationFacade.authentication).thenReturn(authentication)
-      whenever(authentication.authorities).thenReturn(listOf(SimpleGrantedAuthority("ROLE_BOB"), SimpleGrantedAuthority("ROLE_JOE_FRED")))
-      assertThat(userService.myRoles()).containsOnly(ExternalUserRole("BOB"), ExternalUserRole("JOE_FRED"))
-    }
-
-    @Test
-    fun myRoles_noRoles(): Unit = runBlocking {
-      whenever(authenticationFacade.authentication).thenReturn(authentication)
-      whenever(authentication.authorities).thenReturn(emptyList())
-      assertThat(userService.myRoles()).isEmpty()
     }
   }
 }
